% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_2081_2100.R
\docType{data}
\name{cr_inm_cm4_8_ssp126_2081_2100}
\alias{cr_inm_cm4_8_ssp126_2081_2100}
\alias{cr_inm_cm4_8_ssp245_2081_2100}
\alias{cr_inm_cm4_8_ssp370_2081_2100}
\alias{cr_inm_cm4_8_ssp585_2081_2100}
\alias{cr_mpi_esm1_2_hr_ssp126_2081_2100}
\alias{cr_mpi_esm1_2_hr_ssp245_2081_2100}
\alias{cr_mpi_esm1_2_hr_ssp370_2081_2100}
\alias{cr_mpi_esm1_2_hr_ssp585_2081_2100}
\alias{cr_ukesm1_0_ll_ssp126_2081_2100}
\alias{cr_ukesm1_0_ll_ssp245_2081_2100}
\alias{cr_ukesm1_0_ll_ssp370_2081_2100}
\alias{cr_ukesm1_0_ll_ssp585_2081_2100}
\title{Bioclimatic variables downscaled and calibrated CMIP6 climate data for projected future climates of continental Costa Rica (2081 - 2100)}
\format{
A stars object (EPSG:4326)
}
\source{
The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
cr_inm_cm4_8_ssp126_2081_2100
cr_inm_cm4_8_ssp245_2081_2100
cr_inm_cm4_8_ssp370_2081_2100
cr_inm_cm4_8_ssp585_2081_2100
cr_mpi_esm1_2_hr_ssp126_2081_2100
cr_mpi_esm1_2_hr_ssp245_2081_2100
cr_mpi_esm1_2_hr_ssp370_2081_2100
cr_mpi_esm1_2_hr_ssp585_2081_2100
cr_ukesm1_0_ll_ssp126_2081_2100
cr_ukesm1_0_ll_ssp245_2081_2100
cr_ukesm1_0_ll_ssp370_2081_2100
cr_ukesm1_0_ll_ssp585_2081_2100
}
\description{
Includes 3 GCMs (General Circulation Models), INM_CM4_8, MPI-ESM1-2-HR and UKESM1-0-LL, and four SSPs (Shared Socioeconomic Pathways), ssp126, ssp245, ssp370 and ssp585.

Resolution 0.5 minutes of a degree (~1km).
}
\keyword{datasets}
