% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcc-package.R
\docType{package}
\name{crcc-package}
\alias{crcc}
\alias{crcc-package}
\title{crcc: crcc: Costa Rica climate change}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains geospatial data of Costa Rica to do climate change analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://manuelspinola.github.io/crcc/}
}

}
\author{
\strong{Maintainer}: Manuel Spínola \email{mspinola10@gmail.com} (\href{https://orcid.org/0000-0002-7839-1908}{ORCID})

}
\keyword{internal}
