% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcc-package.R
\docType{package}
\name{crcc-package}
\alias{crcc}
\alias{crcc-package}
\title{crcc: crcc: geospatial data of Costa Rica for climate change analysis}
\description{
Contains geospatial data of Costa Rica to do climate change analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://manuelspinola.github.io/crcc/}
}

}
\keyword{internal}
